typing and ts changes
